+++ src/sections/AboutMe/AboutMe.module.css
@@ -116,17 +116,24 @@
 .codeSnippetContainer > div {
   display: flex;
   flex-direction: column;
-  justify-content: space-between;
+  justify-content: space-around;
+  gap: 20px;
 }
 
 .codeSnippetContainer pre {
-  max-height: 450px;
+  max-height: 420px;
   border-radius: 10px;
 }
 
 .windowControllerContainer {
   display: flex;
-  justify-content: flex-start;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.windowControllerContainer > *:last-child * {
+  font-size: 2rem;
+  color: var(--primary);
 }
 
 .windowControllerContainer > div {
@@ -151,4 +158,10 @@
 .hovered svg,
 .clicked svg {
   fill: var(--background);
+}
+
+video {
+  width: 100%;
+  height: 100%;
+  border-radius: 10px;
 }
